requirements.txt
# This file may be used to create an environment using:
# $ pip install -r <this file>
# platform: win-64
# environment: python 3.7
absl-py==0.12.0
asgiref==3.3.4
astor==0.8.1
astroid==2.3.3
astunparse==1.6.3
cached-property==1.5.2
certifi==2021.5.30
click==8.0.1
colorama==0.4.3
dnspython==1.16.0
eventlet==0.31.0
Flask==2.0.1
gast==0.3.3
google-pasta==0.2.0
greenlet==1.1.0
grpcio==1.38.0
h5py==2.10.0
image==1.5.33
importlib-metadata==4.5.0
isort==4.3.21
itsdangerous==2.0.1
Jinja2==3.0.1
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.4.3
Markdown==3.3.4
MarkupSafe==2.0.1
mccabe==0.6.1
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
protobuf==3.17.2
pylint==2.4.4
python-dateutil==2.8.1
python-engineio==3.13.0
python-socketio==4.6.1
pytz==2021.1
PyYAML==5.4.1
scipy==1.6.3
six==1.16.0
sqlparse==0.4.1
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
typed-ast==1.4.1
typing-extensions==3.10.0.0
Werkzeug==2.0.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.4.1
